apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.24.1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
openai==1.3.7
supabase==1.2.0
structlog==23.2.0
python-multipart==0.0.6
python-dotenv==1.1.1 